Run go generate and fix linter comments
