+++ src/mapper/pkg/resolvers/schema.helpers.resolvers.go
@@ -1,1 +1,39 @@
 package resolvers
+
+import (
+	"fmt"
+	"github.com/otterize/intents-operator/src/shared/telemetries/telemetriesgql"
+	"github.com/otterize/intents-operator/src/shared/telemetries/telemetrysender"
+	"github.com/otterize/network-mapper/src/mapper/pkg/graph/model"
+	"github.com/sirupsen/logrus"
+)
+
+type SourceType string
+
+const (
+	SourceTypeDNSCapture  SourceType = "Capture"
+	SourceTypeSocketScan  SourceType = "SocketScan"
+	SourceTypeKafkaMapper SourceType = "KafkaMapper"
+	SourceTypeIstio       SourceType = "Istio"
+)
+
+func updateTelemetriesCounters(sourceType SourceType, intent model.Intent) {
+	clientKey := telemetrysender.Anonymize(fmt.Sprintf("%s/%s", intent.Client.Namespace, intent.Client.Name))
+	serverKey := telemetrysender.Anonymize(fmt.Sprintf("%s/%s", intent.Server.Namespace, intent.Server.Name))
+	intentKey := telemetrysender.Anonymize(fmt.Sprintf("%s-%s", clientKey, serverKey))
+
+	telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeIntentsDiscovered, intentKey)
+	telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeServiceDiscovered, clientKey)
+	telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeServiceDiscovered, serverKey)
+
+	logrus.Infof("Discovered intent %s, %s, %s", clientKey, serverKey, intentKey)
+	if sourceType == SourceTypeDNSCapture {
+		telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeIntentsDiscoveredCapture, intentKey)
+	} else if sourceType == SourceTypeSocketScan {
+		telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeIntentsDiscoveredSocketScan, intentKey)
+	} else if sourceType == SourceTypeKafkaMapper {
+		telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeIntentsDiscoveredKafka, intentKey)
+	} else if sourceType == SourceTypeIstio {
+		telemetrysender.IncrementUniqueCounterNetworkMapper(telemetriesgql.EventTypeIntentsDiscoveredIstio, intentKey)
+	}
+}
